+++ CHANGELOG.md
@@ -9,14 +9,14 @@ and this project adheres to [Forge Semantic Versioning](https://mcforge.readthed
 - Nothing new.
 
 ## [1.18.2-3.2.0.0] - 2022/06/07
-- Fixed [1.18.2-3.2.0.0-beta1] double vertical slabs bug.
+- Fixed [1.18.2-3.2.0.0-beta1] double vertical slabs bugs.
 - Changed project structure to prepare a Fabric port.
 - Fixed slime and honey vertical slabs rendering.
-- Added support for falling blocks.
 
 ## [1.18.2-3.2.0.0-beta1] - 2022/06/04
 - Added double vertical slabs.
 - BUG: Double Vertical Slabs that only have a referredSlabState (so no referredBlockState) don't render correctly.
+- BUG: Double Vertical Slabs that should emit light don't emit it.
 - BUG: (from previous versions) honey and slime vertical slabs don't render correctly.
 
 ## [1.18.2-3.1.0.1] - 2022/05/23